# chapter-6/final-lab/jobposting-back/react-openapi-generator/swagger.yaml
openapi: 3.0.1
info:
  title: Jobposting API
  description: Jobposting API documentation
  termsOfService: ''
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/applicants/{applicantId}:
    put:
      tags:
        - applicant-resource
      operationId: updateApplicant
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Applicant'
    patch:
      tags:
        - applicant-resource
      operationId: partialUpdateApplicant
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/Applicant'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Applicant'
...
  /api/applicants:
    get:
      tags:
        - applicant-resource
      operationId: getAllApplicants
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Applicant'
    post:
      tags:
        - applicant-resource
      operationId: createApplicant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Applicant'
...
  /api/applicants/{id}:
    get:
      tags:
        - applicant-resource
      operationId: getApplicant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Applicant'
    delete:
      tags:
        - applicant-resource
      operationId: deleteApplicant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
...
components:
  schemas:
    Applicant:
      type: object
      properties:
        applicantId:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        phone:
          type: string
...